handle TT in UCI non functional change signature = 16052283

+++ types.h
@@ -96,7 +96,7 @@ inline int opp_color(int color)
 	return color ^ BLACK;
 }
 
-typedef uint32_t Key;
+typedef uint64_t Key;
 typedef uint64_t Bitboard;
 
 #define INF		32767
